578: * split the new huge SurveyEditorPage in two by moving the creatiion of the SurveyCreator object in a new function * in the initialSurveyFormValues we fix the type of expired to enforce a date type (fixes error when opening a survey with an expiration time)
